Makefile.am: enable extra warnings in a non-fatal way to aid developers
The -Wno-error lines are actually TODOs, because almost all could be
solved via well designed clean modifications.
However, fixing all of these is a huge undertaking and is out of scope
for now.

In the near future, we should probably grep for these and make CI fail
if a new commit introduces new warnings. At a minimum, the count of
warnings could be asserted to not increase for any given pull request.

Signed-off-by: bkil-syslogng <tamas.nagy@balabit.com>